[hyper-provider] 1.support pull image with RegistryAuth 2.optimize code
